+++ app/models/surveys/survey.rb
@@ -27,6 +27,7 @@ class Survey < ActiveRecord::Base
   has_many :survey_question_orders, -> { order :question_number }
   has_many :questions, through: :survey_question_orders
   has_many :encounters, -> { where deleted: false }
+  has_many :survey_user_types, -> { where deleted: false }
   has_many :survey_answer_frequencies
   has_many :reports